feat: adds preview for settings
